+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New #186: Add `iconCustomEmojiId` and `style` fields to `InlineKeyboardButton` type.
 - New #186: Add `ChatOwnerLeft` and `ChatOwnerChanged` types.
 - New #186: Add `chatOwnerLeft` and `chatOwnerChanged` fields to `Message` type.
+- New #186: Add `SetMyProfilePhoto` method.
 
 ## 0.14 February 7, 2026